Half Done Exercise 70 - Equivalent Binary Trees
Implemented the Walk function to traverse the binary trees and return
the result through a channel. However, unless the channel is closed by
the goroutine (in this case, the Walk function), the runtime freaks out
that the goroutine is deadlocked. So I had to make another Walk function
called WalkSubtree that actually does the tree walking. The original
Walk function just calls the WalkSubtree function and closes the
channel. It was impossible to make the Walk function recursive and close
the channel inside it without making it overly complex.